project/app/requirements-hw.txt
streamlit
pillow
numpy
scipy
pandas
st-pages
tensorflow
keras
streamlit-drawable-canvas
opencv-python
joblib
scikit-learn
######## AGREGAR TUS LIBRERIAS EN REQUIREMENTS-HW ########

######## AGREGAR TUS LIBRERIAS EN REQUIREMENTS-HW ########
